Updates to make template parsing a utility function

+++ scrub/module_helper.py
@@ -2,6 +2,7 @@ import sys
 import os
 import importlib
 import shutil
+import logging
 import argparse
 from scrub.utils.filtering import do_filtering
 from scrub.utils import scrub_utilities
@@ -15,21 +16,21 @@ def parse_arguments():
 
     # Add parser arguments
     parser.add_argument('--config', default='./scrub.cfg')
-    parser.add_argument('--module', required=True)
+    parser.add_argument('--tool', required=True)
 
     # Parse the arguments
     args = vars(parser.parse_args(sys.argv[2:]))
 
     # Run analysis
-    main(args['module'], args['config'])
+    main(args['tool'], args['config'])
 
 
-def main(scrub_module, conf_file='/.scrub.cfg'):
+def main(tool_name, conf_file='/.scrub.cfg'):
     """
     This function runs a single analysis module, while preserving existing analysis results.
 
     Inputs:
-        --module: Tool import location of the form scrub.tools.<tool>.do_<tool> [string]
+        --tool: Name of tool that needs to be run [string]
         --config: Absolute path to the SCRUB configuration file to be used [string]
     """
 
@@ -38,35 +39,74 @@ def main(scrub_module, conf_file='/.scrub.cfg'):
 
     # Initialize the SCRUB storage directory
     scrub_utilities.initialize_storage_dir(scrub_conf_data)
+    scrub_path = os.path.dirname(os.path.realpath(__file__))
 
-    try:
+    # Find the template of interest
+    template_path = (scrub_path + '/tools/templates/' + scrub_conf_data.get('source_lang') + '/' + tool_name.lower() + '.template')
 
-        # Import the module
-        module_object = importlib.import_module(scrub_module)
+    # Attempt analysis if the template exists
+    if os.path.exists(template_path):
+        try:
+            # Create the log file
+            analysis_log_file = scrub_conf_data.get('scrub_log_dir') + '/' + tool_name + '.log'
+            scrub_utilities.create_logger(analysis_log_file)
 
-        # Call the analysis
-        tool_status = getattr(module_object, "run_analysis")(scrub_conf_data, True)
+            # Print a status message
+            logging.info('  Parsing ' + tool_name + ' template file...')
 
-        # Run filtering, if necessary
-        if (scrub_module.startswith('scrub.tool')) and (tool_status == 0):
-            # Filter and distribute the results
-            do_filtering.run_analysis(scrub_conf_data)
+            # Remove the analysis directory if it exists
+            tool_analysis_dir = os.path.normpath(scrub_conf_data.get('scrub_analysis_dir') + '/' + tool_name +
+                                                 '_analysis')
+            if os.path.exists(tool_analysis_dir):
+                shutil.rmtree(tool_analysis_dir)
 
-    finally:
-        # Move the results back with the source code if necessary
-        if scrub_conf_data.get('scrub_working_dir') != scrub_conf_data.get('scrub_analysis_dir'):
-            # Move every item in the directory
-            for item in os.listdir(scrub_conf_data.get('scrub_working_dir')):
-                # Remove the item if it exists
-                if os.path.exists(scrub_conf_data.get('scrub_analysis_dir') + '/' + item):
-                    if os.path.isfile(scrub_conf_data.get('scrub_analysis_dir') + '/' + item):
-                        os.remove(scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
-                    else:
-                        shutil.rmtree(scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
+            # Create the analysis directory
+            scrub_conf_data.update({'tool_analysis_dir': tool_analysis_dir})
+            os.mkdir(tool_analysis_dir)
 
-                # Move the item
-                shutil.move(scrub_conf_data.get('scrub_working_dir') + '/' + item,
-                            scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
+            # Create the analysis template
+            analysis_script = scrub_utilities.parse_template(template_path, tool_name, scrub_conf_data)
 
-            # Remove the working directory
-            shutil.rmtree(scrub_conf_data.get('scrub_working_dir'))
+            try:
+                # Execute the analysis
+                scrub_utilities.execute_command(analysis_script, os.environ.copy())
+
+                # Filter and distribute the results
+                do_filtering.run_analysis(scrub_conf_data)
+
+            except scrub_utilities.CommandExecutionError:
+                logging.warning(tool_name + ' analysis could not be performed.')
+
+                # Print the exception traceback
+                logging.warning(traceback.format_exc())
+
+            finally:
+                # Close the logger
+                logging.getLogger().handlers = []
+
+
+        finally:
+            # Move the results back with the source code if necessary
+            if scrub_conf_data.get('scrub_working_dir') != scrub_conf_data.get('scrub_analysis_dir'):
+                # Move every item in the directory
+                for item in os.listdir(scrub_conf_data.get('scrub_working_dir')):
+                    # Remove the item if it exists
+                    if os.path.exists(scrub_conf_data.get('scrub_analysis_dir') + '/' + item):
+                        if os.path.isfile(scrub_conf_data.get('scrub_analysis_dir') + '/' + item):
+                            os.remove(scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
+                        else:
+                            shutil.rmtree(scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
+
+                    # Move the item
+                    shutil.move(scrub_conf_data.get('scrub_working_dir') + '/' + item,
+                                scrub_conf_data.get('scrub_analysis_dir') + '/' + item)
+
+                # Remove the working directory
+                shutil.rmtree(scrub_conf_data.get('scrub_working_dir'))
+
+    elif tool_name == 'filter':
+        # Filter and distribute the results
+        do_filtering.run_analysis(scrub_conf_data)
+
+    else:
+        sys.exit('ERROR: Template does not exist.')
